requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.3
pytz==2021.1
sqlparse==0.4.1
